+++ packages/app/src/MainUI.tsx
@@ -1,26 +1,16 @@
 import { useState } from 'react'
-import { useMountReport } from '@masknet/web3-hooks-base'
-import { EventID } from '@masknet/web3-telemetry/types'
-import { DisableShadowRootContext, MaskLightTheme } from '@masknet/theme'
-import { PageUIProvider } from '@masknet/shared'
 import { SortDropdown } from './components/SortDropdown.js'
 import { StickySearchHeader } from './components/StickySearchBar.js'
 import { SidebarForDesktop } from './components/SidebarForDesktop.js'
 import { SidebarForMobile } from './components/SidebarForMobile.js'
 import { ActivityFeed } from './components/ActivityFeed.js'
-import { DecryptUI } from './main/index.js'
+import { DecryptMessage } from './main/index.js'
 
-function useTheme() {
-    return MaskLightTheme
-}
-
-function MainUI() {
+export function MainUI() {
     const [sidebarOpen, setSidebarOpen] = useState(false)
-
     return (
         <div className="bg-zinc-900 h-full">
             <SidebarForMobile sidebarOpen={sidebarOpen} setSidebarOpen={setSidebarOpen} />
-
             <SidebarForDesktop />
 
             <div className="xl:pl-72">
@@ -35,7 +25,7 @@ function MainUI() {
 
                     <div className="bg-white p-5">
                         <div className="border pt-3 rounded-lg">
-                            <DecryptUI />
+                            <DecryptMessage />
                         </div>
                     </div>
                 </main>
@@ -43,15 +33,5 @@ function MainUI() {
                 <ActivityFeed />
             </div>
         </div>
-    )
-}
-
-export function App() {
-    useMountReport(EventID.AccessPopups)
-
-    return (
-        <DisableShadowRootContext.Provider value>
-            {PageUIProvider(useTheme, <MainUI />)}
-        </DisableShadowRootContext.Provider>
     )
 }
