+++ packages/app/src/index.tsx
@@ -1,15 +1,29 @@
 import './styles/index.css'
-import './locales/legacy.js'
+import './locales/init.js'
 
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { App } from './App.js'
+import { PageUIProvider } from '@masknet/shared'
+import { DisableShadowRootContext, MaskLightTheme } from '@masknet/theme'
+import { useMountReport } from '@masknet/web3-hooks-base'
+import { EventID } from '@masknet/web3-telemetry/types'
+import { MainUI } from './MainUI.js'
 
 const root = document.createElement('main')
 document.body.appendChild(root)
 
 createRoot(root).render(
     <StrictMode>
-        <App />
+        <DisableShadowRootContext.Provider value>
+            <App />
+        </DisableShadowRootContext.Provider>
     </StrictMode>,
 )
+function App() {
+    useMountReport(EventID.AccessPopups)
+
+    return PageUIProvider(useTheme, <MainUI />)
+}
+function useTheme() {
+    return MaskLightTheme
+}
