+++ packages/app/src/main/TypedMessageRender/registry.ts
@@ -1,0 +1,3 @@
+import { createTypedMessageRenderRegistry } from '@masknet/typed-message-react'
+
+export const registry = createTypedMessageRenderRegistry()
