+++ packages/app/src/plugin-host/enable.ts
@@ -1,6 +1,5 @@
 import './register.js'
 
-import { noop } from 'lodash-es'
 import { Emitter } from '@servie/events'
 import { CurrentSNSNetwork, SNSAdaptorContextRef, startPluginSNSAdaptor } from '@masknet/plugin-infra/content-script'
 import type { Plugin } from '@masknet/plugin-infra/content-script'
@@ -14,7 +13,6 @@ import {
     ValueRefWithReady,
 } from '@masknet/shared-base'
 import { setupReactShadowRootEnvironment } from '@masknet/theme'
-import type { UnboundedRegistry } from '@dimensiondev/holoflows-kit'
 import { ThemeMode, FontSize } from '@masknet/web3-shared-base'
 import { addListener } from './message.js'
 import { PluginWorker } from './rpc.js'
@@ -46,27 +44,6 @@ async function reject(): Promise<never> {
     throw new Error('Not implemented')
 }
 
-const asyncNoop = async () => {}
-const emptyEventRegistry: UnboundedRegistry<any> = {
-    send: noop,
-    off: noop,
-    on: () => noop,
-    sendByBroadcast: noop,
-    sendToAll: noop,
-    sendToBackgroundPage: noop,
-    pause: () => asyncNoop,
-    sendToContentScripts: noop,
-    sendToFocusedPage: noop,
-    sendToLocal: noop,
-    sendToVisiblePages: noop,
-    bind: () => ({
-        off: noop,
-        on: () => noop,
-        pause: () => asyncNoop,
-        send: noop,
-    }),
-    async *[Symbol.asyncIterator]() {},
-}
 const emptyValueRef = new ValueRefWithReady<any>()
 
 startPluginSNSAdaptor(CurrentSNSNetwork.__SPA__, {
